Move existing tag related notice to right place.
