+++ assets/js/modules/adsense/components/settings/SettingsForm.js
@@ -28,10 +28,7 @@ import { Fragment } from '@wordpress/element';
 import Data from 'googlesitekit-data';
 import { ProgressBar } from 'googlesitekit-components';
 import { MODULES_ADSENSE } from '../../datastore/constants';
-import {
-	parseAccountID,
-	parseAccountIDFromExistingTag,
-} from '../../util/parsing';
+import { parseAccountID } from '../../util/parsing';
 import {
 	ErrorNotices,
 	UseSnippetSwitch,
@@ -42,7 +39,6 @@ import WebStoriesAdUnitSelect from '../common/WebStoriesAdUnitSelect';
 import Link from '../../../../components/Link';
 import { CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
 import AdBlockingRecoveryCTA from '../common/AdBlockingRecoveryCTA';
-import SettingsNotice from '../../../../components/SettingsNotice/SettingsNotice';
 import { useFeature } from '../../../../hooks/useFeature';
 const { useSelect } = Data;
 
@@ -52,9 +48,6 @@ export default function SettingsForm() {
 	const webStoriesActive = useSelect( ( select ) =>
 		select( CORE_SITE ).isWebStoriesActive()
 	);
-	const accountID = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getAccountID()
-	);
 	const clientID = useSelect( ( select ) =>
 		select( MODULES_ADSENSE ).getClientID()
 	);
@@ -63,9 +56,6 @@ export default function SettingsForm() {
 	);
 	const hasResolvedGetExistingTag = useSelect( ( select ) =>
 		select( MODULES_ADSENSE ).hasFinishedResolution( 'getExistingTag' )
-	);
-	const existingAdBlockingRecoveryTag = useSelect( ( select ) =>
-		select( MODULES_ADSENSE ).getExistingAdBlockingRecoveryTag()
 	);
 	const hasResolvedGetExistingAdBlockingRecoveryTag = useSelect( ( select ) =>
 		select( MODULES_ADSENSE ).hasFinishedResolution(
@@ -110,26 +100,6 @@ export default function SettingsForm() {
 		);
 	}
 
-	let existingAdBlockingRecoveryTagMessage;
-	if (
-		existingAdBlockingRecoveryTag &&
-		existingAdBlockingRecoveryTag === accountID
-	) {
-		existingAdBlockingRecoveryTagMessage = __(
-			'You’ve already got an Ad Blocking Recovery code on your site. We recommend you use Site Kit to manage this to get the most out of AdSense.',
-			'google-site-kit'
-		);
-	} else if ( existingAdBlockingRecoveryTag ) {
-		existingAdBlockingRecoveryTagMessage = sprintf(
-			/* translators: %s: account ID */
-			__(
-				'Site Kit detected Ad Blocking Recovery code for a different account %s on your site. For a better ad blocking recovery experience, you should remove Ad Blocking Recovery code that’s not linked to this AdSense account.',
-				'google-site-kit'
-			),
-			parseAccountIDFromExistingTag( existingAdBlockingRecoveryTag )
-		);
-	}
-
 	const supportURL =
 		'https://support.google.com/adsense/answer/10175505#create-an-ad-unit-for-web-stories';
 
@@ -165,12 +135,6 @@ export default function SettingsForm() {
 			) }
 
 			<AutoAdExclusionSwitches />
-
-			{ existingAdBlockingRecoveryTag && (
-				<SettingsNotice
-					notice={ existingAdBlockingRecoveryTagMessage }
-				/>
-			) }
 
 			{ adBlockerDetectionEnabled && (
 				<Fragment>
